Add solution for blog list app, exs 4.15-4.23

+++ controllers/blogs.js
@@ -1,8 +1,9 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  let blogs = await Blog.find({})
+  let blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -11,9 +12,16 @@ blogsRouter.get('/:id', async (request, response) => {
   response.json(blog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === user.id) {
+    await Blog.findByIdAndDelete(request.params.id)
+    return response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'only creator of the blog can delete it' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -26,17 +34,23 @@ blogsRouter.put('/:id', async (request, response) => {
   response.json(updatedBlog)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const user = request.user
   const body = request.body
-  body.likes = body.likes === undefined ? 0 : body.likes
 
   if (!body.title || !body.url) {
-    response.status(400).end()
-    return
+    return response.status(400).end()
   }
 
-  const blog = new Blog(body)
+  body.likes = body.likes === undefined ? 0 : body.likes
+
+  const blog = new Blog({
+    ...body,
+    user: user.id
+  })
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
 
   response.status(201).json(savedBlog)
 })
